requirements.txt
flask==3.1.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==21.2.0
flask-cors==4.0.0
